Move EntityPersister to drivers directory.
